+++ tests/testdata/collection/sketch-2.js
@@ -1,0 +1,8 @@
+function setup() {
+    createCanvas(windowWidth, windowHeight);
+    text('Sketch #2', 10, 10);
+}
+
+function draw() {
+    circle(mouseX, mouseY, 20);
+}