+++ tests/testdata/descriptions/multi-line-block-description.js
@@ -1,0 +1,10 @@
+/**
+ * Description: sketch
+ * description
+ *
+ * Additional text that does not belong to the Description
+ */
+
+function setup() {
+    createCanvas(100, 100);
+}