+++ tests/testdata/descriptions/single-line-block-description.js
@@ -1,0 +1,9 @@
+/**
+ * Description: sketch description
+ *
+ * Additional text that does not belong to the Description
+ */
+
+function setup() {
+    createCanvas(100, 100);
+}