+++ tests/testdata/descriptions/single-line-description.js
@@ -1,0 +1,7 @@
+// Description: sketch description
+//
+// Additional text that does not belong to the description
+
+function setup() {
+    createCanvas(100, 100);
+}