+++ tests/testdata/html-includes/sketch.js
@@ -1,0 +1,8 @@
+function preload() {
+    loadImage("cat.png");
+    loadJson("data.json");
+}
+
+function setup() {
+    createCanvas(windowWidth, windowHeight);
+}